feat: 삭제 관련 로직 이동
